requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-cosmos
apache-airflow-providers-common-sql
apache-airflow-providers-apache-kafka
apache-airflow-providers-postgres
apache-airflow-providers-oracle
apache-airflow-providers-amazon
apache-airflow-providers-apache-iceberg
apache-airflow-providers-trino
airflow-provider-duckdb
dremio-simple-query
python-dotenv
oracledb
nodejs
